+++ src/Os.ts
@@ -1,0 +1,16 @@
+function getOs() {
+  const platform = process.platform;
+
+  switch (platform) {
+    case "darwin":
+      return "macOS";
+    case "win32":
+      return "Windows";
+    case "linux":
+      return "Linux";
+    default:
+      return platform;
+  }
+}
+
+export const os = getOs();