+++ src/extension.ts
@@ -4,12 +4,14 @@ import { fileNestingDataProvider } from "./FileNestingDataProvider";
 import { fileNestingTreeViewExplorer } from "./FileNestingTreeViewExplorer";
 // import { fileNestingDecoratorProvider } from "./FileNestingDecoratorProvider";
 import { createFileNestingCommands } from "./FileNestingCommands";
+import { initMixpanel } from "./commands/analytics";
 
 export function activate(context: vscode.ExtensionContext) {
   console.log(
     'Congratulations, your extension "file-nesting-explorer" is now active!'
   );
 
+  initMixpanel();
   fileNestingDataProvider.setContext(context);
   fileNestingTreeViewExplorer.setContext(context);