+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec2.0.0.html).
 
+## [0.3.2] - 2025-11-21
+
+### Added
+
+- Mixpanel analytics to track usage of main extension commands (New Nested File, ...)
+
 ## [0.3.1] - 2025-11-21
 
 ### Fixed